[REFACTOR] Put all error_handlers in one namespace
